Add files via upload
CLI bash script has been debugged and runs all utilities sucessfully. You should be able to replace instances of $pythoncall with whatever python call variable your system uses. The script would not run sucessfully with a defined pythoncall variable in the script, but does with user-defined input for reasons I am not certain of. The batch script has not yet been debugged and trialed on an actual Windows DPU connected to the eVOLVER.